federation/requirements.txt
# federation service dependencies
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
